requirements.txt
absl-py==2.0.0
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
basemap==1.3.8
basemap-data==1.3.2
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==2.2.1
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflate64==1.0.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
ml-collections==0.1.1
ml-dtypes==0.2.0
multivolumefile==0.2.3
nest-asyncio==1.5.8
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.2
pycryptodomex==3.19.0
Pygments==2.16.1
pyparsing==3.1.1
pyppmd==1.1.0
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-ternary==1.0.8
pytorch-fid==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.15.9
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.22.0
termcolor==2.3.0
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
